requirements.txt
build==1.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
distutils-pytest==0.2.1
Django==2.2.13
django-bootstrap-icons==0.8.7
django-bootstrap3==11.0.0
django-constance==2.4.0
django-debug-toolbar==1.11
django-jfu==2.0.9
django-model-utils==3.1.2
django-picklefield==3.0.1
django-su==0.8.0
django-utils-six==2.0
django-versatileimagefield==1.11
django-wysiwyg==0.7.0
fpdf==1.7.2
idna==3.6
iniconfig==2.0.0
install==1.3.5
libmagic==1.0
packaging==24.0
pillow==10.3.0
pip-tools==7.4.1
pluggy==1.4.0
pyproject_hooks==1.1.0
pytest==8.1.1
python-magic==0.4.27
pytz==2024.1
requests==2.31.0
setuptools==69.2.0
six==1.11.0
sorl-thumbnail==12.3
sqlparse==0.2.4
Unidecode==1.0.22
urllib3==2.2.1
wheel==0.43.0
